ROM, RAM, Decoder
- Dec8080 provides address space decoding
  one edge per KiB of memory space for ROM
  one edge per 4 KiB of memory space for RAM
  one edge per I/O port
  TODO: separate I/O input from output space
- Rom8080 provides read-only storage
  each instance is a 1024x8 bit ROM
  TODO: provide initial load of ROM data
- Ram8080 provides read-write storage
- Sys8080 now has 8 ROMs and 16 RAMs

Also:
- unnamed union of unnamed struct not needed for 8080 STATUS byte
- prefer Octal to Hex for 8080 related things
- make more hardware-like things ALL UPPERCASE
- remove erroneous semicolons in #define lines in 8080_status.h
- fix definition of Word: needs to be 16 bit for 8080

+++ inc/word.hh
@@ -5,4 +5,4 @@
 // change for simulated 32-bit and 64-bit systems.
 
 #include <cstdint>
-typedef uint32_t Word;
+typedef uint16_t Word;
